Handle null scenarios for the byline
