+++ src/app/pages/ArticlePage/ArticlePage.jsx
@@ -59,6 +59,7 @@ import {
   getArticleSection,
   getMentions,
   getLang,
+  getAuthorTwitterHandle,
 } from '#lib/utilities/parseAssetData';
 import filterForBlockType from '#lib/utilities/blockHandlers';
 import RelatedTopics from '#containers/RelatedTopics';
@@ -138,47 +139,9 @@ const ArticlePage = ({ pageData, mostReadEndpointOverride }) => {
   const startsWithHeading = propEq('type', 'headline')(blocks[0] || {});
   const hasByline = blocks.find(block => block.type === 'byline');
 
-  const articleAuthorTwitterHandle = (contentBlocks => {
-    if (!hasByline) return null;
-
-    const [firstByline, ...otherBylines] = contentBlocks.filter(
-      block => block.type === 'byline',
-    );
-    const [firstContributor, ...otherContributors] = pathOr(
-      null,
-      ['model', 'blocks'],
-      firstByline,
-    );
-    const contributorBlocks = pathOr(
-      null,
-      ['model', 'blocks'],
-      firstContributor,
-    );
-    const contributorLink = contributorBlocks.find(
-      block => block.type === 'link',
-    );
-    const contributorHandle = pathOr(
-      null,
-      [
-        'model',
-        'blocks',
-        0,
-        'model',
-        'blocks',
-        0,
-        'model',
-        'blocks',
-        0,
-        'model',
-        'blocks',
-        0,
-        'model',
-        'text',
-      ],
-      contributorLink,
-    );
-    return contributorHandle;
-  })(blocks);
+  const articleAuthorTwitterHandle = hasByline
+    ? getAuthorTwitterHandle(blocks)
+    : null;
 
   const componentsToRender = {
     visuallyHiddenHeadline,
